refactor(fs): move object storage code to object subpackage
